requirements.txt
flask
flask-cors
pdfplumber
python-docx
spacy
google-generativeai
python-dotenv
